requirements.txt
azure-devops==7.1.0b4
requests==2.31.0
python-dotenv==1.0.0
openai==1.35.7
pydantic==2.7.4
pytest==7.4.4
pytest-mock==3.11.1
pytest-asyncio==0.21.1
